remove the internals package
