--- internals/path-to-regexp/package.json
@@ -1,22 +1,0 @@
-{
-  "private": true,
-  "name": "@vercel-internals/path-to-regexp",
-  "version": "1.0.0",
-  "types": "dist/index.d.ts",
-  "main": "dist/index.js",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc -p tsconfig.json"
-  },
-  "dependencies": {
-    "path-to-regexp-updated": "npm:path-to-regexp@6.3.0",
-    "path-to-regexp": "6.1.0"
-  },
-  "devDependencies": {
-    "@vercel-internals/tsconfig": "1.0.0",
-    "@vercel/style-guide": "4.0.2",
-    "typescript": "4.9.4"
-  }
-}
