--- internals/path-to-regexp/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@vercel-internals/tsconfig",
-  "compilerOptions": {
-    "outDir": "./dist"
-  },
-  "include": ["src/**/*.ts"]
-}
