+++ packages/node/package.json
@@ -36,6 +36,8 @@
     "esbuild": "0.14.47",
     "etag": "1.8.1",
     "node-fetch": "2.6.9",
+    "path-to-regexp-updated": "npm:path-to-regexp@6.3.0",
+    "path-to-regexp": "6.1.0",
     "ts-morph": "12.0.0",
     "ts-node": "10.9.1",
     "typescript": "4.9.5",
@@ -52,7 +54,6 @@
     "@types/etag": "1.8.0",
     "@types/jest": "29.5.0",
     "@vercel/functions": "workspace:*",
-    "@vercel-internals/path-to-regexp": "1.0.0",
     "@vitest/expect": "1.4.0",
     "content-type": "1.0.5",
     "cookie": "0.7.0",
