+++ packages/remix/package.json
@@ -26,6 +26,8 @@
     "@vercel/error-utils": "2.0.3",
     "@vercel/nft": "0.27.10",
     "@vercel/static-config": "3.0.0",
+    "path-to-regexp-updated": "npm:path-to-regexp@6.3.0",
+    "path-to-regexp": "6.1.0",
     "ts-morph": "12.0.0"
   },
   "devDependencies": {
@@ -33,7 +35,6 @@
     "@types/node": "14.18.33",
     "@types/semver": "7.3.13",
     "@vercel/build-utils": "9.2.1",
-    "@vercel-internals/path-to-regexp": "1.0.0",
     "glob": "10.3.16",
     "jest-junit": "16.0.0",
     "semver": "7.5.2",
